requirements.txt
Flask==3.1.0
Pillow==11.1.0
timm==1.0.15
torch==2.6.0
torchvision==0.21.0
